remove audio control overlap from endless scroll
The audioplayer plugin loads every time more results are loaded on to
the page. As a result the appending html from the script overlaps with
the previously loaded results.

It might be best to get rid of the plugin altogether, but for the time
being, we'll address the problem by hiding the extra html with css so
that it isn't visible to the user. Also, a promise has been added to
end of the plugin so that the function removes all overlapped instances
of audioplayer class divs.

Unfortunately, the hiding and removal of tags causes interuptions to
the playback.